pyproject.toml
[project]
name = "v1"
version = "0.1.0"
description = "A production-grade GPT-2 clone."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mlx>=0.26.2",
    "tiktoken>=0.9.0",
    "torch>=2.7.1",
    "transformers>=4.53.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.2",
    "pytest>=8.3.2",
    "wandb>=0.17.5",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "black>=24.4.2",
    "ruff>=0.5.5",
    "mypy>=1.11.0",
]

[tool.setuptools.packages.find]
where = ["src"]
